pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ebus-simulator"
version = "0.1.0"
description = "Standalone linear EBUS simulator with localizer, physics, and review workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "nibabel>=5.3",
  "numpy>=2.2",
  "Pillow>=11.0",
  "PyYAML>=6.0",
  "scipy>=1.15",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]
ui = [
  "PySide6>=6.8",
]

[project.scripts]
validate-case = "ebus_simulator.cli:main"
generate-poses = "ebus_simulator.pose_cli:main"
render-preset = "ebus_simulator.render_cli:main"
render-all-presets = "ebus_simulator.render_all_cli:main"
review-presets = "ebus_simulator.review_cli:main"
compare-review-bundles = "ebus_simulator.review_compare_cli:main"
analyze-render-consistency = "ebus_simulator.consistency_cli:main"
launch-app = "ebus_simulator.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--capture=sys"
markers = [
  "unit: fast synthetic or helper-level tests that do not require the dataset-backed render path",
  "integration: slower dataset-backed or end-to-end rendering, review, and app tests",
]
